Fixed compilation errors on Windows
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@404 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/powertabdocument/complexsymbolarray.h
@@ -53,7 +53,7 @@ namespace ComplexSymbols
     size_t getComplexSymbolCount(const std::array<uint32_t, N>& symbolArray)
     {
         return  symbolArray.size() -
-                std::count(symbolArray.begin(), symbolArray.end(), notUsed);
+                std::count(symbolArray.begin(), symbolArray.end(), static_cast<uint32_t>(notUsed));
     }
 
     /// Clears the contents of the symbol array (sets all elements to "not used")
